requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
filelock==3.0.12
fugashi==1.1.1
huggingface-hub==0.0.12
idna==3.2
importlib-metadata==4.8.1
ipadic==1.0.0
joblib==1.0.1
kiwisolver==1.3.2
matplotlib==3.4.3
mecab-python==1.0.0
mecab-python3==1.0.4
numpy==1.21.2
openTSNE==0.6.0
packaging==21.0
pandas==1.3.3
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.21
requests==2.26.0
sacremoses==0.0.45
scikit-learn==0.22.2.post1
scipy==1.7.1
seaborn==0.11.2
sentencepiece==0.1.96
six==1.16.0
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.0
tornado==6.1
tqdm==4.62.1
transformers==4.9.2
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.6
zipp==3.6.0
